🔒Security : 블로그 게시판 인가 관련 수정

+++ src/main/java/com/portalSite/blog/service/BlogBoardService.java
@@ -20,9 +20,12 @@ public class BlogBoardService {
     private final BlogBoardRepository blogBoardRepository;
     private final BlogRepository blogRepository;
 
-    public BlogBoardResponse saveBlogBoard(CreateBlogBoardRequest request, Long blogId) {
+    public BlogBoardResponse saveBlogBoard(CreateBlogBoardRequest request, Long blogId, Long memberId) {
         Blog blog = blogRepository.findById(blogId)
             .orElseThrow(() -> new RuntimeException("존재하지 않는 블로그입니다."));
+        if (!blog.getMember().getId().equals(memberId)) {
+            throw new RuntimeException("다른 사람의 블로그입니다.");
+        }
         BlogBoard blogBoard = BlogBoard.of(blog, request.getCategory());
 
         return BlogBoardResponse.from(blogBoardRepository.save(blogBoard));
@@ -43,9 +46,13 @@ public class BlogBoardService {
         return BlogBoardResponse.from(blogBoardRepository.save(blogBoard));
     }
 
-    public void deleteBlogBoard(Long categoryId) {
+    public void deleteBlogBoard(Long categoryId, Long memberId) {
         BlogBoard blogBoard = blogBoardRepository.findById(categoryId)
             .orElseThrow(() -> new RuntimeException("존재하지 않는 게시판입니다."));
+
+        if (!blogBoard.getBlog().getMember().getId().equals(memberId)) {
+            throw new RuntimeException("다른 사람의 블로그입니다.");
+        }
 
         blogBoardRepository.delete(blogBoard);
     }
